Wrote initial version of rock-paper-scissors
The html file to host the js file has been written. Within the rps.js
we have the game which takes either a full word or first initial of the
hand the player wants to play. If the hand isn't valid then we return
a message letting the user know. Then the script generates a random hand
to be played by the computer and evaluates who won.

 Changes to be committed:
	new file:   index.html
	new file:   rps.js